% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_utils.R
\docType{data}
\name{basis_function_registry}
\alias{basis_function_registry}
\title{Registry of known basis functions and their reproducible attributes}
\format{
An object of class \code{list} of length 3.
}
\usage{
basis_function_registry
}
\description{
Registry of known basis functions and their reproducible attributes
}
\keyword{internal}
