% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_utils.R
\name{bt_if_needed}
\alias{bt_if_needed}
\title{Conditionally add backticks based on syntax validity}
\usage{
bt_if_needed(x, data = NULL)
}
\arguments{
\item{x}{A character string}

\item{data}{Optional data frame to check if x exists as a column name}
}
\value{
The name, potentially backticked
}
\description{
Add backticks only if the name is not a valid R syntactic name.
}
\keyword{internal}
