% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdata.R
\name{complete_data}
\alias{complete_data}
\title{Filter out entities with too few observations}
\usage{
complete_data(data, formula = NULL, vars = NULL, min.waves = "all")
}
\arguments{
\item{data}{A \code{\link[=panel_data]{panel_data()}} frame.}

\item{formula}{A formula, like the one you'll be using to specify your model.}

\item{vars}{As an alternative to formula, a vector of variable names.}

\item{min.waves}{What is the minimum number of observations to be kept?
Default is \code{"all"}, but it can be any number.}
}
\value{
A \code{panel_data} frame.
}
\description{
This function allows you to define a minimum number of
waves/periods and exclude all individuals with fewer observations than
that.
}
