% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_utils.R
\name{detect_matrix_terms}
\alias{detect_matrix_terms}
\title{Check if any terms in a formula are matrix-returning}
\usage{
detect_matrix_terms(terms, data)
}
\arguments{
\item{terms}{Character vector of formula terms}

\item{data}{Data frame to evaluate against}
}
\value{
Logical vector
}
\description{
Check if any terms in a formula are matrix-returning
}
\keyword{internal}
