% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_utils.R
\name{expand_basis_columns}
\alias{expand_basis_columns}
\title{Expand a basis matrix into individual columns in a data frame}
\usage{
expand_basis_columns(data, mat, colnames)
}
\arguments{
\item{data}{Data frame to add columns to}

\item{mat}{Matrix to expand}

\item{colnames}{Character vector of column names}
}
\value{
Data frame with added columns
}
\description{
Expand a basis matrix into individual columns in a data frame
}
\keyword{internal}
