% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_formula.R
\name{extract_ranef_vars}
\alias{extract_ranef_vars}
\title{Extract variables from random effects terms}
\usage{
extract_ranef_vars(ranefs, data)
}
\arguments{
\item{ranefs}{Character vector of random effects terms}

\item{data}{Data frame to check for variable existence}
}
\value{
Character vector of variable names
}
\description{
Helper to extract variable names from random effects specifications
}
\keyword{internal}
