% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_formula.R
\name{get_interactions.WBFormula}
\alias{get_interactions.WBFormula}
\title{Get all interaction labels from WBFormula}
\usage{
get_interactions.WBFormula(x, type = c("all", "within", "cross", "between"))
}
\arguments{
\item{x}{A WBFormula object}

\item{type}{One of "all", "within", "cross", or "between"}
}
\value{
Character vector of interaction labels
}
\description{
Get all interaction labels from WBFormula
}
\keyword{internal}
