% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_formula.R
\name{get_meanvar}
\alias{get_meanvar}
\title{Get mean variable name for a term}
\usage{
get_meanvar(x, term)
}
\arguments{
\item{x}{A WBFormula object}

\item{term}{The term to look up}
}
\value{
The mean variable name, or NULL if not found
}
\description{
Get mean variable name for a term
}
\keyword{internal}
