% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdata.R
\name{get_wave}
\alias{get_wave}
\alias{get_id}
\title{Retrieve panel_data metadata}
\usage{
get_wave(data)

get_id(data)
}
\arguments{
\item{data}{A \code{panel_data} frame}
}
\value{
A \code{panel_data} frame
}
\description{
\code{get_id} and \code{get_wave} are extractor functions that can be
used to retrieve the names of the id and wave variables of a \code{panel_data}
frame.
}
