% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_formula.R
\name{has_interactions}
\alias{has_interactions}
\title{Check if WBFormula has interactions}
\usage{
has_interactions(x)
}
\arguments{
\item{x}{A WBFormula object}
}
\value{
Logical indicating if any interactions are present
}
\description{
Check if WBFormula has interactions
}
\keyword{internal}
