% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fixef.wbm}
\alias{fixef.wbm}
\alias{ranef.wbm}
\alias{isGLMM.wbm}
\alias{isLMM.wbm}
\alias{isNLMM.wbm}
\title{Helpers functions for interfacing with wbm objects}
\usage{
fixef.wbm(object, add.dropped = FALSE, ...)

ranef.wbm(object, condVar = FALSE, drop = FALSE, whichel = NULL,
  postVar = FALSE, ...)

isGLMM.wbm(x, ...)

isLMM.wbm(x, ...)

isNLMM.wbm(x, ...)
}
\arguments{
\item{object}{A \code{wbm} model object}

\item{add.dropped}{for models with rank-deficient design matrix,
reconstitute the full-length parameter vector by adding NA values in
appropriate locations?}

\item{...}{some methods for these generic functions require additional
arguments}

\item{condVar}{an optional logical argument indicating if the conditional
variance-covariance matrices of the random effects should be added as an
attribute.}

\item{drop}{should components of the return value that would be data frames
with a single column, usually a column called \code{(Intercept)}, be returned as
named vectors instead?}

\item{whichel}{character vector of names of grouping factors for which the
random effects should be returned.}

\item{postVar}{a (deprecated) synonym for condVar}

\item{x}{A \code{wbm} model object}
}
\description{
Helpers functions for interfacing with wbm objects
}
