% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_utils.R
\name{is_known_basis_fn}
\alias{is_known_basis_fn}
\title{Check if a function is a known basis function}
\usage{
is_known_basis_fn(fn_name)
}
\arguments{
\item{fn_name}{Character string of function name}
}
\value{
Logical
}
\description{
Check if a function is a known basis function
}
\keyword{internal}
