% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_utils.R
\name{is_matrix_term}
\alias{is_matrix_term}
\title{Check if a term returns a matrix when evaluated}
\usage{
is_matrix_term(term, data)
}
\arguments{
\item{term}{Character string representing a formula term}

\item{data}{Data frame to evaluate against (should be ungrouped)}
}
\value{
Logical indicating if the term returns a matrix
}
\description{
Check if a term returns a matrix when evaluated
}
\keyword{internal}
