% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_data.R
\name{is_panel}
\alias{is_panel}
\title{Check if object is panel_data}
\usage{
is_panel(x)
}
\arguments{
\item{x}{Any object.}
}
\description{
This is a convenience function that checks whether an object
is a \code{panel_data} object.
}
\examples{
 data("WageData")
 is_panel(WageData) # FALSE
 wages <- panel_data(WageData, id = id, wave = t)
 is_panel(wages) # TRUE
}
