% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_formula.R
\name{is_varying_term}
\alias{is_varying_term}
\title{Check if a variable is time-varying in WBFormula}
\usage{
is_varying_term(x, var)
}
\arguments{
\item{x}{A WBFormula object}

\item{var}{Variable name to check}
}
\value{
Logical
}
\description{
Check if a variable is time-varying in WBFormula
}
\keyword{internal}
