% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_utils.R
\name{make_interaction_config}
\alias{make_interaction_config}
\title{Create InteractionConfig from wbm() arguments}
\usage{
make_interaction_config(interaction.style, model, detrend)
}
\arguments{
\item{interaction.style}{The interaction.style argument}

\item{model}{The model argument}

\item{detrend}{The detrend argument}
}
\value{
An InteractionConfig object
}
\description{
Factory function to create InteractionConfig from the
arguments passed to wbm() or wbgee().
}
\keyword{internal}
