% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nloptwrap_alt}
\alias{nloptwrap_alt}
\title{Alternate optimizer for \code{wbm} and other models}
\usage{
nloptwrap_alt(fn, par, lower, upper, control = list(), ...)
}
\arguments{
\item{fn}{A function to be minimized (or maximized), with first
argument the vector of parameters over which minimization is
to take place.  It should return a scalar result.}

\item{par}{a vector of initial values for the parameters for which
optimal values are to be found. Names on the elements of this
vector are preserved and used in the results data frame.}

\item{lower}{Bounds on the variables for methods such as ‘"L-BFGS-B"’
that can handle box (or bounds) constraints.}

\item{upper}{Bounds on the variables for methods such as ‘"L-BFGS-B"’
that can handle box (or bounds) constraints.}

\item{control}{A list of control parameters.}

\item{...}{Further arguments passed to \code{\link[nloptr:nloptr]{nloptr::nloptr()}}}
}
\description{
Alternate optimizer for \code{wbm} and other models
}
\details{
This optimizer is exported for use in \code{\link[=wbm]{wbm()}} to improve
speed and reliability of optimization.
}
