% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-panel_data.R
\name{panel_data-vctrs}
\alias{panel_data-vctrs}
\alias{vec_restore.panel_data}
\alias{vec_proxy.panel_data}
\alias{vec_ptype2.panel_data}
\alias{vec_ptype2.panel_data.panel_data}
\alias{vec_ptype2.panel_data.data.frame}
\alias{vec_ptype2.data.frame.panel_data}
\alias{vec_ptype2.panel_data.tbl_df}
\alias{vec_ptype2.tbl_df.panel_data}
\alias{vec_cast.panel_data}
\alias{vec_cast.panel_data.panel_data}
\alias{vec_cast.panel_data.data.frame}
\alias{vec_cast.panel_data.tbl_df}
\alias{vec_cast.data.frame.panel_data}
\alias{vec_cast.tbl_df.panel_data}
\title{Internal vctrs methods}
\usage{
\method{vec_restore}{panel_data}(x, to, ...)

\method{vec_proxy}{panel_data}(x, ...)

\method{vec_ptype2}{panel_data}(x, y, ...)

\method{vec_ptype2}{panel_data.panel_data}(x, y, ...)

\method{vec_ptype2}{panel_data.data.frame}(x, y, ...)

\method{vec_ptype2}{data.frame.panel_data}(x, y, ...)

\method{vec_ptype2}{panel_data.tbl_df}(x, y, ...)

\method{vec_ptype2}{tbl_df.panel_data}(x, y, ...)

\method{vec_cast}{panel_data}(x, to, ...)

\method{vec_cast}{panel_data.panel_data}(x, to, ...)

\method{vec_cast}{panel_data.data.frame}(x, to, ...)

\method{vec_cast}{panel_data.tbl_df}(x, to, ...)

\method{vec_cast}{data.frame.panel_data}(x, to, ...)

\method{vec_cast}{tbl_df.panel_data}(x, to, ...)
}
\description{
These methods are the extensions that allow panel_data objects to
work with vctrs and modern dplyr/tidyr operations.
}
\keyword{internal}
