% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_formula.R
\name{print.WBFormula}
\alias{print.WBFormula}
\title{Print method for WBFormula}
\usage{
\method{print}{WBFormula}(x, ...)
}
\arguments{
\item{x}{A WBFormula object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for WBFormula
}
