% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_utils.R
\name{reconstruct_basis_call}
\alias{reconstruct_basis_call}
\title{Reconstruct a basis function call with a modified variable}
\usage{
reconstruct_basis_call(term, new_var, attrs = NULL)
}
\arguments{
\item{term}{Original term like "ns(age, df=3)"}

\item{new_var}{New variable expression like "age - imean(age)"}

\item{attrs}{Optional list of attributes to add as arguments}
}
\value{
Character string of the new call
}
\description{
Reconstruct a basis function call with a modified variable
}
\keyword{internal}
