% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_utils.R
\name{should_demean_ints}
\alias{should_demean_ints}
\title{Determine if interactions should be de-meaned}
\usage{
should_demean_ints(config)
}
\arguments{
\item{config}{An InteractionConfig object}
}
\value{
Logical indicating whether to demean interactions
}
\description{
Based on the interaction configuration, determine whether
interaction terms should have their means subtracted.
}
\keyword{internal}
