% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teen_poverty.R
\docType{data}
\name{teen_poverty}
\alias{teen_poverty}
\title{National Longitudinal Survey of Youth teenage women poverty data}
\format{A data frame with 1141 rows and 28 variables:
\describe{
\item{id}{Unique identifier for the respondent}
\item{age}{Age at first interview}
\item{black}{1 if subject is black, 0 if not}
\item{pov1}{1 if subject is in poverty, 0 if not, at time 1}
\item{pov2}{1 if subject is in poverty, 0 if not, at time 2}
\item{pov3}{1 if subject is in poverty, 0 if not, at time 3}
\item{pov4}{1 if subject is in poverty, 0 if not, at time 4}
\item{pov5}{1 if subject is in poverty, 0 if not, at time 5}
\item{mother1}{1 if subject has had a child, 0 if not, at time 1}
\item{mother2}{1 if subject has had a child, 0 if not, at time 2}
\item{mother3}{1 if subject has had a child, 0 if not, at time 3}
\item{mother4}{1 if subject has had a child, 0 if not, at time 4}
\item{mother5}{1 if subject has had a child, 0 if not, at time 5}
\item{spouse1}{1 if subject lives with a spouse, 0 if not, at time 1}
\item{spouse2}{1 if subject lives with a spouse, 0 if not, at time 2}
\item{spouse3}{1 if subject lives with a spouse, 0 if not, at time 3}
\item{spouse4}{1 if subject lives with a spouse, 0 if not, at time 4}
\item{spouse5}{1 if subject lives with a spouse, 0 if not, at time 5}
\item{inschool1}{1 if subject is in school, 0 if not, at time 1}
\item{inschool2}{1 if subject is in school, 0 if not, at time 2}
\item{inschool3}{1 if subject is in school, 0 if not, at time 3}
\item{inschool4}{1 if subject is in school, 0 if not, at time 4}
\item{inschool5}{1 if subject is in school, 0 if not, at time 5}
\item{hours1}{Hours worked during the week of the survey, at time 1}
\item{hours2}{Hours worked during the week of the survey, at time 2}
\item{hours3}{Hours worked during the week of the survey, at time 3}
\item{hours4}{Hours worked during the week of the survey, at time 4}
\item{hours5}{Hours worked during the week of the survey, at time 5}
}}
\source{
These data originate with the U.S. Department of Labor. The particular
subset used here come from Paul Allison via Statistical Horizons:
\url{https://statisticalhorizons.com/wp-content/uploads/teenpov.dta}
}
\usage{
teen_poverty
}
\description{
These data come from the years 1979-1983 in the National Longitudinal
Survey of Youth, with information about 1141 teenage women. These data are in
the "wide" format for demonstration purposes.
}
\keyword{datasets}
