% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_utils.R
\name{un_bt}
\alias{un_bt}
\title{Remove backticks from names}
\usage{
un_bt(x)
}
\arguments{
\item{x}{A character vector potentially containing backticks}
}
\value{
A character vector with backticks removed
}
\description{
Remove all backticks from variable names. Useful for cleaning
names after formula parsing.
}
\keyword{internal}
