% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_utils.R
\name{update_pf_for_matrix_terms}
\alias{update_pf_for_matrix_terms}
\title{Update parsed formula object for matrix terms}
\usage{
update_pf_for_matrix_terms(pf, expanded)
}
\arguments{
\item{pf}{WBFormula object}

\item{expanded}{Result from expand_matrix_terms_in_data}
}
\value{
Updated WBFormula object
}
\description{
Updates the v_info and other fields in pf to reflect the expanded
matrix term columns.
}
\keyword{internal}
