% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_utils.R
\name{use_old_style_ints}
\alias{use_old_style_ints}
\title{Determine if "old-style" interaction processing is needed}
\usage{
use_old_style_ints(config)
}
\arguments{
\item{config}{An InteractionConfig object}
}
\value{
Logical indicating whether to use old-style processing
}
\description{
Old-style processing creates interaction terms BEFORE
demeaning the constituent variables.
}
\keyword{internal}
