% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_formula.R
\name{WBFormula_from_parser}
\alias{WBFormula_from_parser}
\title{Create WBFormula from parser output (for migration)}
\usage{
WBFormula_from_parser(pf, formula, dv)
}
\arguments{
\item{pf}{List output from wb_formula_parser()}

\item{formula}{The original Formula object}

\item{dv}{The dependent variable name}
}
\value{
A WBFormula object
}
\description{
Create WBFormula from parser output (for migration)
}
\keyword{internal}
