% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_wbMod.R
\docType{class}
\name{wbm-class}
\alias{wbm-class}
\title{Within-Between Model (\code{wbm}) class}
\description{
Models fit using \code{\link[=wbm]{wbm()}} return values of this class, which
inherits from \code{\link[lme4]{merMod-class}}.
}
\section{Slots}{

\describe{
\item{\code{call_info}}{A list of metadata about the arguments used.}

\item{\code{call}}{The actual function call.}

\item{\code{summ}}{The \code{\link[jtools:summ]{jtools::summ()}} object returned from calling it on the
\code{merMod} object.}

\item{\code{summ_atts}}{The attributes of the \code{summ} object.}

\item{\code{orig_data}}{The data provided to the \code{data} argument in the function
call.}
}}

