% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_stan.R
\name{wbm_stan}
\alias{wbm_stan}
\title{Bayesian estimation of within-between models}
\usage{
wbm_stan(formula, data, id = NULL, wave = NULL, model = "w-b",
  use.wave = FALSE, wave.factor = FALSE, min.waves = 2,
  model.cor = FALSE, family = gaussian, fit_model = TRUE, chains = 3,
  iter = 2000, scale = FALSE, save_ranef = FALSE, weights = NULL, ...)
}
\arguments{
\item{formula}{Model formula. See details for crucial
info on \code{panelr}'s formula syntax.}

\item{data}{The data, either a \code{panel_data} object or \code{data.frame}.}

\item{id}{If \code{data} is not a \code{panel_data} object, then the name of the
individual id column. Otherwise, leave as NULL, the default.}

\item{wave}{If \code{data} is not a \code{panel_data} object, then the name of the
panel wave column. Otherwise, leave as NULL, the default.}

\item{model}{One of \code{"w-b"}, \code{"within"}, \code{"between"},
\code{"contextual"}, or \code{"stability"}. See details for more on these options.}

\item{use.wave}{Should the wave be included as a predictor? Default is
FALSE.}

\item{wave.factor}{Should the wave variable be treated as an unordered
factor instead of continuous? Default is FALSE.}

\item{min.waves}{What is the minimum number of waves an individual must
have participated in to be included in the analysis? Default is \code{2} and
any valid number is accepted. \code{"all"} is also acceptable if you want to
include only complete panelists.}

\item{model.cor}{Do you want to model residual autocorrelation?
This is often appropriate for linear models (\code{family = gaussian}).
Default is FALSE to be consistent with \code{\link[=wbm]{wbm()}}, reduce
runtime, and avoid warnings for non-linear models.}

\item{family}{Use this to specify GLM link families. Default is \code{gaussian},
the linear model.}

\item{fit_model}{Fit the model? Default is TRUE. If FALSE, only the model
code is returned.}

\item{chains}{How many Markov chains should be used? Default is 3, to leave
you with one unused thread if you're on a typical dual-core machine.}

\item{iter}{How many iterations, including warmup? Default is 2000, leaving
1000 per chain after warmup. For some models and data, you may need quite
a few more.}

\item{scale}{Standardize predictors? This can speed up model fit. Default
is FALSE.}

\item{save_ranef}{Save random effect estimates? This can be crucial for
predicting from the model and for certain post-estimation procedures.
On the other hand, it drastically increases the size of the resulting
model. Default is FALSE.}

\item{weights}{If using weights, either the name of the column in the data
that contains the weights or a vector of the weights.}

\item{...}{Additional arguments passed on to \code{\link[brms:brm]{brms::brm()}}. This can include
specification of priors.}
}
\value{
A \code{wbm_stan} object, which is a list containing a \code{model} object
with the \code{brm} model and a \code{stan_code} object with the model code.

If \code{fit_model = FALSE}, insead a list is returned containing a \code{stan_code}
object and a \code{stan_data} object, leaving you with the tools you need to
run the model yourself using \code{rstan}.
}
\description{
A near-equivalent of \code{\link[=wbm]{wbm()}} that instead uses Stan,
via \pkg{rstan} and \pkg{brms}.
}
\details{
See \code{\link[=wbm]{wbm()}} for details on the formula syntax, model types,
and some other stuff.
}
\examples{
\dontrun{
if(interactive()){
 data("WageData")
 wages <- panel_data(WageData, id = id, wave = t)
 model <- wbm_stan(lwage ~ lag(union) + wks | blk + fem | blk * lag(union),
           data = wages, chains = 1, iter = 2000)
 summary(model)
 }
}
}
\seealso{
\code{\link[=wbm]{wbm()}}
}
\author{
Jacob A. Long
}
