% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdata.R
\name{widen_panel}
\alias{widen_panel}
\title{Convert long panel data to wide format}
\usage{
widen_panel(data, separator = "_")
}
\arguments{
\item{data}{The \code{panel_data} frame.}

\item{separator}{When the variables are labeled with the wave number,
what should separate the variable name and wave number? By default,
it is "_". In other words, a variable named \code{var} will be
\code{var_1}, \code{var_2}, and so on in the wide data frame.}
}
\value{
A data.frame with 1 row per respondent.
}
\description{
This function takes \code{\link[=panel_data]{panel_data()}} objects as input as converts
them to wide format for use in SEM and other situations when such a format
is needed.
}
\details{
This is a wrapper for \code{\link[stats:reshape]{stats::reshape()}}, which is renowned for being
pretty confusing to use. This function automatically detects which of the
variables vary over time and which don't, not appending wave information
to constants.
}
\examples{

wages <- panel_data(WageData, id = id, wave = t)
wide_wages <- widen_panel(wages)

}
\seealso{
\code{\link[stats]{reshape}}
}
