% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_data.R
\name{get_wave}
\alias{get_wave}
\alias{get_id}
\alias{get_periods}
\title{Retrieve panel_data metadata}
\usage{
get_wave(data)

get_id(data)

get_periods(data)
}
\arguments{
\item{data}{A \code{panel_data} frame}
}
\value{
A \code{panel_data} frame
}
\description{
\code{get_id()}, \code{get_wave()}, and \code{get_periods()} are extractor
functions that can be used to retrieve the names of the id and wave
variables or time periods of a \code{panel_data} frame.
}
\examples{

data("WageData")
wages <- panel_data(WageData, id = id, wave = t)
get_wave(wages)
get_id(wages)
get_periods(wages)

}
