% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_methods.R, R/pdata.R
\name{as_panel_data}
\alias{as_panel_data}
\alias{as_panel_data.default}
\alias{as_panel_data.pdata.frame}
\alias{as_panel}
\alias{panel_data}
\title{Create panel data frames}
\usage{
as_panel_data(x, ...)

\method{as_panel_data}{default}(x, id = id, wave = wave, ...)

\method{as_panel_data}{pdata.frame}(x, ...)

as_panel(x, ...)

panel_data(data, id = id, wave = wave, ...)
}
\arguments{
\item{...}{Attributes for adding onto this method. See
\code{\link[tibble:new_tibble]{tibble::new_tibble()}} for a run-through of the logic.}

\item{id}{The name of the column (unquoted) that identifies
participants/entities. A new column will be created called \code{id},
overwriting any column that already has that name.}

\item{wave}{The name of the column (unquoted) that identifies
waves or periods. A new column will be created called \code{wave},
overwriting any column that already has that name.}

\item{data}{A data frame.}
}
\value{
A \code{panel_data} object.
}
\description{
Format your data for use with \pkg{panelr}.
}
\examples{
data("WageData")
wages <- panel_data(WageData, id = id, wave = t)

}
